Refactor: removed logging from util
